+++ src/modules/images/images.module.ts
@@ -8,5 +8,6 @@ import { ImageService } from './images.service';
   imports: [TypeOrmModule.forFeature([Image])],
   controllers: [ImageController],
   providers: [ImageService],
+  exports: [ImageService],
 })
 export class ImagesModule {}
