+++ src/modules/tasks/tasks.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { ImagesModule } from '../images/images.module';
+import { TasksService } from './tasks.service';
+
+@Module({
+  imports: [ImagesModule],
+  providers: [TasksService],
+})
+export class TasksModule {}
